pyproject.toml
[project]
name = "time-mcp"
version = "0.1.0"
description = "时间处理 MCP 工具 - 提供时间计算、时区转换等功能"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
keywords = ["mcp", "time", "timezone", "fastmcp", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = ["fastmcp>=2.0.0", "python-dateutil>=2.9.0", "pytz>=2024.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/time_mcp"]
